+++ tests/measures.R
@@ -47,23 +47,23 @@ model <- SISe(u0      = data.frame(S = 99, I = 1),
               end_t4  = 365,
               epsilon = 0)
 
-res <- assertError(trajectory(model, compartments = "S", as.is = TRUE))
+res <- assertError(trajectory(model, compartments = "S", format = "matrix"))
 check_error(res, "Please run the model first, the trajectory is empty.")
 
-res <- assertError(trajectory(model, compartments = "I", as.is = TRUE))
+res <- assertError(trajectory(model, compartments = "I", format = "matrix"))
 check_error(res, "Please run the model first, the trajectory is empty.")
 
 result <- run(model)
 result
 
-res <- assertError(trajectory(result, I ~ S, as.is = TRUE))
+res <- assertError(trajectory(result, I ~ S, format = "matrix"))
 check_error(res, "Invalid formula specification of 'compartments'.")
 
 stopifnot(identical(
-    length(trajectory(result, compartments = "S", as.is = TRUE)),
+    length(trajectory(result, compartments = "S", format = "matrix")),
     1001L))
 stopifnot(identical(
-    length(trajectory(result, compartments = "I", as.is = TRUE)),
+    length(trajectory(result, compartments = "I", format = "matrix")),
     1001L))
 
 p <- prevalence(result, I ~ S + I, format = "matrix")
@@ -80,10 +80,10 @@ if (SimInf:::have_openmp()) {
     result
 
     stopifnot(identical(
-        length(trajectory(result, compartments = "S", as.is = TRUE)),
+        length(trajectory(result, compartments = "S", format = "matrix")),
         1001L))
     stopifnot(identical(
-        length(trajectory(result, compartments = "I", as.is = TRUE)),
+        length(trajectory(result, compartments = "I", format = "matrix")),
         1001L))
 
     p <- prevalence(result, I ~ S + I, format = "matrix")
@@ -113,20 +113,20 @@ model <- SISe_sp(u0       = data.frame(S = 99, I = 1),
                  coupling = 0,
                  distance = distance_matrix(1, 1, 1))
 
-res <- assertError(trajectory(model, compartments = "S", as.is = TRUE))
+res <- assertError(trajectory(model, compartments = "S", format = "matrix"))
 check_error(res, "Please run the model first, the trajectory is empty.")
 
-res <- assertError(trajectory(model, compartments = "I", as.is = TRUE))
+res <- assertError(trajectory(model, compartments = "I", format = "matrix"))
 check_error(res, "Please run the model first, the trajectory is empty.")
 
 result <- run(model)
 result
 
 stopifnot(identical(
-    length(trajectory(result, compartments = "S", as.is = TRUE)),
+    length(trajectory(result, compartments = "S", format = "matrix")),
     1001L))
 stopifnot(identical(
-    length(trajectory(result, compartments = "I", as.is = TRUE)),
+    length(trajectory(result, compartments = "I", format = "matrix")),
     1001L))
 
 p <- prevalence(result, I ~ S + I, format = "matrix")
@@ -143,10 +143,10 @@ if (SimInf:::have_openmp()) {
     result
 
     stopifnot(identical(
-        length(trajectory(result, compartments = "S", as.is = TRUE)),
+        length(trajectory(result, compartments = "S", format = "matrix")),
         1001L))
     stopifnot(identical(
-        length(trajectory(result, compartments = "I", as.is = TRUE)),
+        length(trajectory(result, compartments = "I", format = "matrix")),
         1001L))
 
     p <- prevalence(result, I ~ S + I, format = "matrix")
@@ -182,20 +182,20 @@ model <- SISe3(u0        = u0,
                end_t4    = 365,
                epsilon   = 0.000011)
 
-res <- assertError(trajectory(model, compartments = "S_1", as.is = TRUE))
+res <- assertError(trajectory(model, compartments = "S_1", format = "matrix"))
 check_error(res, "Please run the model first, the trajectory is empty.")
 
-res <- assertError(trajectory(model, compartments = "I_1", as.is = TRUE))
+res <- assertError(trajectory(model, compartments = "I_1", format = "matrix"))
 check_error(res, "Please run the model first, the trajectory is empty.")
 
 result <- run(model)
 result
 
 stopifnot(identical(
-    length(trajectory(result, compartments = "S_1", as.is = TRUE)),
+    length(trajectory(result, compartments = "S_1", format = "matrix")),
     10000L))
 stopifnot(identical(
-    length(trajectory(result, compartments = "I_1", as.is = TRUE)),
+    length(trajectory(result, compartments = "I_1", format = "matrix")),
     10000L))
 
 p <- prevalence(result, I_1 + I_2 + I_3 ~ ., format = "matrix")
@@ -212,10 +212,10 @@ if (SimInf:::have_openmp()) {
     result
 
     stopifnot(identical(
-        length(trajectory(result, compartments = "S_1", as.is = TRUE)),
+        length(trajectory(result, compartments = "S_1", format = "matrix")),
         10000L))
     stopifnot(identical(
-        length(trajectory(result, compartments = "I_1", as.is = TRUE)),
+        length(trajectory(result, compartments = "I_1", format = "matrix")),
         10000L))
 
     p <- prevalence(result, I_1 + I_2 + I_3 ~ ., format = "matrix")
@@ -254,20 +254,20 @@ model <- SISe3_sp(u0        = u0,
                   coupling = 0,
                   distance = distance_matrix(1:10, 1:10, 1))
 
-res <- assertError(trajectory(model, compartments = "S_1", as.is = TRUE))
+res <- assertError(trajectory(model, compartments = "S_1", format = "matrix"))
 check_error(res, "Please run the model first, the trajectory is empty.")
 
-res <- assertError(trajectory(model, compartments = "I_1", as.is = TRUE))
+res <- assertError(trajectory(model, compartments = "I_1", format = "matrix"))
 check_error(res, "Please run the model first, the trajectory is empty.")
 
 result <- run(model)
 result
 
 stopifnot(identical(
-    length(trajectory(result, compartments = "S_1", as.is = TRUE)),
+    length(trajectory(result, compartments = "S_1", format = "matrix")),
     10000L))
 stopifnot(identical(
-    length(trajectory(result, compartments = "I_1", as.is = TRUE)),
+    length(trajectory(result, compartments = "I_1", format = "matrix")),
     10000L))
 
 p <- prevalence(result, I_1 + I_2 + I_3 ~ ., format = "matrix")
@@ -284,10 +284,10 @@ if (SimInf:::have_openmp()) {
     result
 
     stopifnot(identical(
-        length(trajectory(result, compartments = "S_1", as.is = TRUE)),
+        length(trajectory(result, compartments = "S_1", format = "matrix")),
         10000L))
     stopifnot(identical(
-        length(trajectory(result, compartments = "I_1", as.is = TRUE)),
+        length(trajectory(result, compartments = "I_1", format = "matrix")),
         10000L))
 
     p <- prevalence(result, I_1 + I_2 + I_3 ~ ., format = "matrix")